ui: move functions to ui handler
